# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/roles": {
      "get": {
        "operationId": "listAdminRoles",
        "parameters": [
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List roles",
        "tags": [
          "admin-roles"
        ]
      },
      "post": {
        "operationId": "createAdminRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleListItemDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create a role",
        "tags": [
          "admin-roles"
        ]
      }
    },
    "/api/v1/admin/roles/permissions": {
      "get": {
        "operationId": "listAdminPermissions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List permissions",
        "tags": [
          "admin-roles"
        ]
      }
    },
    "/api/v1/admin/roles/{roleId}": {
      "patch": {
        "operationId": "updateAdminRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoleListItemDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Update a role",
        "tags": [
          "admin-roles"
        ]
      },
      "delete": {
        "operationId": "deleteAdminRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteRoleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessControlErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete a role",
        "tags": [
          "admin-roles"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog Platform Public API",
    "description": "Public and administration HTTP contract.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "blog_session"
      }
    },
    "schemas": {
...
      "AccessControlErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "USER_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "The requested user was not found."
          },
          "statusCode": {
            "type": "number",
            "example": 404
          }
        },
        "required": [
          "code",
          "message",
          "statusCode"
        ]
      },
...
      "PermissionSummaryDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "code",
          "description",
          "id"
        ]
      },
      "AdminRoleListItemDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isSystem": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionSummaryDto"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userCount": {
            "type": "number"
          }
        },
        "required": [
          "code",
          "createdAt",
          "description",
          "id",
          "isSystem",
          "name",
          "permissions",
          "updatedAt",
          "userCount"
        ]
      },
      "AdminRoleListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminRoleListItemDto"
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 80,
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "permissionIds": {
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "code",
          "name",
          "permissionIds"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 500
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "permissionIds": {
            "maxItems": 200,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "DeleteRoleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}